requirements.txt
actionlib==1.14.0
angles==1.9.13
ansitable==0.10.0
asttokens==2.4.1
backcall==0.2.0
base_local_planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
cfgv==3.4.0
colored==1.4.4
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cras-py-common==2.3.9
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.8
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
einops==0.8.0
executing==2.1.0
filelock==3.15.4
fonttools==4.54.1
fsspec==2024.6.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
h5py==3.11.0
identify==2.5.36
image-geometry==1.16.2
importlib_resources==6.4.5
interactive-markers==1.12.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.4
laser_geometry==1.6.7
# Editable install with no version control (lightglue==0.0)
-e /home/hanglok/Desktop/ur_slam-main/LightGlue
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
message-filters==1.16.0
moveit-commander==1.1.14
moveit-core==1.1.14
moveit-ros-planning-interface==1.1.14
moveit-ros-visualization==1.1.14
mpmath==1.3.0
networkx==3.1
nodeenv==1.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pgraph-python==0.6.2
pickleshare==0.7.5
pillow==10.4.0
pilz-robot-programming==0.5.0
platformdirs==4.2.2
point-cloud-transport==1.0.11
pre-commit==3.5.0
progress==1.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.5
pybullet==3.2.6
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
roboticstoolbox-python==1.1.1
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospy==1.16.0
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rtb-data==1.0.1
rviz==1.14.25
scipy==1.10.1
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
spatialgeometry==1.1.0
spatialmath-python==1.1.10
srdfdom==0.6.4
stack-data==0.6.3
stevedore==5.2.0
swift-sim==1.1.0
sympy==1.12.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
topic-tools==1.16.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
trac_ik_python==1.6.6
traitlets==5.14.3
transforms3d==0.4.2
triton==2.3.1
turtlebot3_teleop==1.2.5
typing_extensions==4.12.2
urdfdom-py==0.4.6
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wcwidth==0.2.13
websockets==12.0
xacro==1.14.18
zipp==3.20.2
